feat: add txType as a separate file under transactions and update receipt.rs accordingly.
