+++ crates/op-rpc-types/src/op/transaction/mod.rs
@@ -1,1 +1,2 @@
 mod receipt;
+mod tx_type;
